requirements.txt
# Core dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0

# HTML parsing (for Zoom email conversion)
beautifulsoup4>=4.12.0

# Google API (Gmail)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# GCP Services (Firestore, Secret Manager)
google-cloud-firestore>=2.13.0
google-cloud-secret-manager>=2.16.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
